+++ src/backend/edusharing/sign-jwt.ts
