+++ src/backend/edusharing/verify-jwt.ts
